+++ src/main/java/roomescape/controller/handler/ControllerAdvice.java
@@ -1,0 +1,25 @@
+package roomescape.controller.handler;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.exception.PastReservationException;
+import roomescape.exception.ReservationExistsException;
+
+@RestControllerAdvice("roomescape.controller")
+public class ControllerAdvice {
+    @ExceptionHandler(PastReservationException.class)
+    public ResponseEntity<String> handlePastReservationException(PastReservationException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(ReservationExistsException.class)
+    public ResponseEntity<String> handleReservationExistsException(ReservationExistsException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+}
