--- src/main/java/roomescape/controller/handler/RestControllerAdvice.java
@@ -1,18 +1,0 @@
-package roomescape.controller.handler;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@org.springframework.web.bind.annotation.RestControllerAdvice
-public class RestControllerAdvice {
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleException(Exception e){
-        return ResponseEntity.internalServerError().body("예기치 못한 오류가 발생했습니다.");
-    }
-}
